# brilirs deps + buffering