--- a/bril-rs/bril2json/src/cli.rs
+++ b/bril-rs/bril2json/src/cli.rs
@@ -4,6 +4,6 @@ use clap::Parser;
 #[clap(about, version, author)] // keeps the cli synced with Cargo.toml
 pub struct Cli {
     /// Flag for whether position information should be included
-    #[clap(short)]
+    #[clap(short, action)]
     pub position: bool,
 }