--- a/bril-rs/brild/src/cli.rs
+++ b/bril-rs/brild/src/cli.rs
@@ -5,9 +5,9 @@ use std::path::PathBuf;
 #[clap(about, version, author)] // keeps the cli synced with Cargo.toml
 pub struct Cli {
     /// The bril file to statically link. stdin is assumed if file is not provided.
-    #[clap(short, long)]
+    #[clap(short, long, action)]
     pub file: Option<String>,
     /// A list of library paths to look for Bril files.
-    #[clap(short, long, multiple_values = true)]
+    #[clap(short, long, action, multiple_values = true)]
     pub libs: Vec<PathBuf>,
 }