--- a/brilirs/Cargo.toml
+++ b/brilirs/Cargo.toml
@@ -13,12 +13,12 @@ keywords = ["compiler", "bril", "interpreter", "data-structures", "language"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [build-dependencies]
-clap         = { version = "3.1", features = ["derive"] }
-clap_complete= "3.1"
+clap         = { version = "3.2", features = ["derive"] }
+clap_complete= "3.2"
 
 [dependencies]
 thiserror    = "1.0"
-clap         = { version = "3.1", features = ["derive"] }
+clap         = { version = "3.2", features = ["derive"] }
 fxhash       = "0.2"
 mimalloc     = "0.1"