--- a/brilirs/src/cli.rs
+++ b/brilirs/src/cli.rs
@@ -5,22 +5,23 @@ use clap::Parser;
 #[clap(allow_hyphen_values(true))]
 pub struct Cli {
   /// Flag to output the total number of dynamic instructions
-  #[clap(short, long)]
+  #[clap(short, long, action)]
   pub profile: bool,
 
   /// The bril file to run. stdin is assumed if file is not provided
-  #[clap(short, long)]
+  #[clap(short, long, action)]
   pub file: Option<String>,
 
   /// Flag to only typecheck/validate the bril program
-  #[clap(short, long)]
+  #[clap(short, long, action)]
   pub check: bool,
 
   /// Flag for when the bril program is in text form
-  #[clap(short, long)]
+  #[clap(short, long, action)]
   pub text: bool,
 
   /// Arguments for the main function
+  #[clap(action)]
   pub args: Vec<String>,
 
   /// This is the original source file that the file input was generated from.
@@ -28,6 +29,6 @@ pub struct Cli {
   /// interpreter is in JSON form with source positions and you what brilirs to
   /// include sections of the source file in its error messages.
   /// If --text/-t is provided, that will be assumed to be the source file if none are provided.
-  #[clap(short, long)]
+  #[clap(short, long, action)]
   pub source: Option<String>,
 }