This pr handles the ever marching forward of dependency versions and switches to using `BufWriter` when writing to `stdout` which, as https://doc.rust-lang.org/std/io/struct.BufWriter.html describes, is much more efficient for many small calls to write.

Looking at this now, I realize that there is maybe one case where the `BufWriter` does not get flushed: when the Bril program raises an `InterpError` and then there is an error during the dropping of `BufWriter` where its contents are not flushed. I think this is both unlikely in practice and an allowable edge case.