@Pat-Lafon 2022-06-23
I probably should have included some numbers to justify making a change for "performance reasons". Here is a silly benchmark of `function_call.bril` except it prints the argument it was passed at the end of the call.

| `brili` | old `brilirs` | new `brilirs` | `brilift -j` | `brilift` aot |
| -- | ---| --- | ----| --- | 
| 14.226 s ±  0.187 s | 2.128 s ±  0.008 s | 658.7 ms ±   3.3 ms | 1.772 s ±  0.022 s | 1.082 s ±  0.024 s |